Support both Live (x64) and EMU (Win32) eqlib builds
Guard the ~12-year RoF2->Live client divergences so the plugin compiles
against both eqlib branches (branch selects client via BuildType.h).

- InventoryData: ItemDefinition combat-mod/heroic-resist fields were removed
  on live; switch to ItemBase::Get*() accessors (emu real value, live 0) via
  new member-fn-pointer aug helpers (AugmentedStatFn/augTotalFn/candFn/wornFn)
- CharData: guard emu-only RaidMember::IsInZone with IS_EMU_CLIENT (live infers
  presence from GetSpawnByName); static_cast HPCurrent (int32->int64 on live)
- PlayerModule: static_cast target HPCurrent for the int64 widening

+++ Modules/PlayerModule.cpp
@@ -233,7 +233,7 @@ void PlayerModule::DrawTarget()
 
 	std::string targetName = mq::IsAnonymized() ? "Target" : myui::TrimName(pTarget->DisplayedName);
 
-	int pct = pTarget->HPCurrent;
+	int pct = static_cast<int>(pTarget->HPCurrent);
 	bool overlayText = m_ctx.UI->Flag(GetName(), "TargetTextOverlay", false);
 
 	if (overlayText)
